pyproject.toml
[project]
name = "bspqa"
version = "0.1.0"
description = "RAG-powered Q&A over Bangko Sentral ng Pilipinas Monetary Policy and Inflation Reports. Ask questions about Philippine inflation, interest rates, and economic outlook, and get cited, sourced answers."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "openai>=1.0.0",
    "opentelemetry-api>=1.27.0",
    "opentelemetry-exporter-otlp-proto-http>=1.27.0",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "opentelemetry-sdk>=1.27.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0.0",
    "qdrant-client>=1.17.1",
    "uvicorn>=0.44.0",
]

[project.optional-dependencies]
ingestion = [
    "markitdown[az-doc-intel,pdf]>=0.1.5",
    "tiktoken>=0.7.0",
]

[dependency-groups]
dev = [
    "mypy>=1.10",
    "pytest>=8.0.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "tiktoken>=0.7.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
